+++ src/main/java/br/nnpe/Numero.java
@@ -4,7 +4,7 @@ public class Numero {
 	private Double numero;
 
 	public Numero(Integer ptsCarreira) {
-		this.numero = new Double(ptsCarreira);
+		this.numero = new Double(ptsCarreira.doubleValue());
 	}
 
 	public Double getNumero() {
